scripts/oreVesting/requirements.txt
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
colander==1.8.3
ecdsa==0.17.0
idna==3.3
iso8601==1.0.2
libeospy==2.1.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
translationstring==1.4
urllib3==1.26.9
